move r and rg listitem to common
