+++ android/app/src/main/java/ly/david/mbjc/ui/release/ReleaseScaffold.kt
@@ -27,13 +27,11 @@ import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.paging.PagingData
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.launch
 import ly.david.musicsearch.core.models.network.MusicBrainzEntity
 import ly.david.mbjc.ui.release.details.ReleaseDetailsScreen
-import ly.david.mbjc.ui.release.stats.ReleaseStatsUi
 import ly.david.mbjc.ui.release.tracks.TracksByReleaseScreen
 import ly.david.musicsearch.core.models.listitem.ListItemModel
 import ly.david.musicsearch.strings.LocalStrings
@@ -228,12 +226,12 @@ internal fun ReleaseScaffold(
 
                 ReleaseTab.STATS -> {
 //                    ReleaseStatsUi(
-////                        releaseId = releaseId,
+// //                        releaseId = releaseId,
 //                        modifier = Modifier
 //                            .padding(innerPadding)
 //                            .fillMaxSize()
 //                            .nestedScroll(scrollBehavior.nestedScrollConnection),
-////                        tabs = ReleaseTab.values().map { it.tab }.toImmutableList(),
+// //                        tabs = ReleaseTab.values().map { it.tab }.toImmutableList(),
 //                    )
                 }
             }
