+++ feature/search/src/commonMain/kotlin/ly/david/musicsearch/feature/search/internal/SearchResultsScreen.kt
@@ -1,5 +1,6 @@
 package ly.david.musicsearch.feature.search.internal
 
+import ReleaseListItem
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyListState
@@ -37,6 +38,7 @@ import ly.david.ui.common.label.LabelListItem
 import ly.david.ui.common.paging.ScreenWithPagingLoadingAndError
 import ly.david.ui.common.place.PlaceListItem
 import ly.david.ui.common.recording.RecordingListItem
+import ly.david.ui.common.releasegroup.ReleaseGroupListItem
 import ly.david.ui.common.series.SeriesListItem
 import ly.david.ui.common.work.WorkListItem
 
@@ -68,26 +70,24 @@ internal fun SearchResultsScreen(
             }
 
             is ReleaseGroupListItemModel -> {
-
                 // TODO: should see album type rather than year
-//                ReleaseGroupListItem(releaseGroup = listItemModel) {
-//                    onItemClick(
-//                        MusicBrainzEntity.RELEASE_GROUP,
-//                        id,
-//                        getNameWithDisambiguation(),
-//                    )
-//                }
+                ReleaseGroupListItem(releaseGroup = listItemModel) {
+                    onItemClick(
+                        MusicBrainzEntity.RELEASE_GROUP,
+                        id,
+                        getNameWithDisambiguation(),
+                    )
+                }
             }
 
             is ReleaseListItemModel -> {
-                // TODO: handle release and rg list items in commonMain
-//                ReleaseListItem(release = listItemModel) {
-//                    onItemClick(
-//                        MusicBrainzEntity.RELEASE,
-//                        id,
-//                        getNameWithDisambiguation(),
-//                    )
-//                }
+                ReleaseListItem(release = listItemModel) {
+                    onItemClick(
+                        MusicBrainzEntity.RELEASE,
+                        id,
+                        getNameWithDisambiguation(),
+                    )
+                }
             }
 
             is RecordingListItemModel -> {
