+++ ui/common-legacy/src/main/java/ly/david/ui/commonlegacy/release/ReleasesListScreen.kt
@@ -1,24 +1,20 @@
 package ly.david.ui.commonlegacy.release
 
+import ReleaseListItem
 import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import app.cash.paging.PagingData
 import app.cash.paging.compose.LazyPagingItems
-import app.cash.paging.compose.collectAsLazyPagingItems
-import kotlinx.coroutines.flow.MutableStateFlow
 import ly.david.musicsearch.core.models.getNameWithDisambiguation
+import ly.david.musicsearch.core.models.listitem.EndOfList.id
 import ly.david.musicsearch.core.models.listitem.ReleaseListItemModel
 import ly.david.musicsearch.core.models.network.MusicBrainzEntity
 import ly.david.ui.common.listitem.SwipeToDeleteListItem
 import ly.david.ui.common.paging.ScreenWithPagingLoadingAndError
-import ly.david.ui.core.preview.DefaultPreviews
-import ly.david.ui.core.theme.PreviewTheme
 import org.koin.androidx.compose.koinViewModel
 import timber.log.Timber
 
@@ -108,18 +104,18 @@ internal fun ReleasesListScreenInternal(
     }
 }
 
-@DefaultPreviews
-@Composable
-internal fun PreviewReleasesListScreen() {
-    PreviewTheme {
-        Surface {
-            val items = MutableStateFlow(
-                PagingData.from(
-                    ReleasePreviewParameterProvider().values.toList(),
-                ),
-            )
-
-            ReleasesListScreenInternal(lazyPagingItems = items.collectAsLazyPagingItems())
-        }
-    }
-}
+// @DefaultPreviews
+// @Composable
+// internal fun PreviewReleasesListScreen() {
+//    PreviewTheme {
+//        Surface {
+//            val items = MutableStateFlow(
+//                PagingData.from(
+//                    ReleasePreviewParameterProvider().values.toList(),
+//                ),
+//            )
+//
+//            ReleasesListScreenInternal(lazyPagingItems = items.collectAsLazyPagingItems())
+//        }
+//    }
+// }
