+++ ui/common-legacy/src/test/java/ly/david/ui/commonlegacy/release/ReleaseListScreenTest.kt
@@ -1,17 +1,12 @@
 package ly.david.ui.commonlegacy.release
 
-import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import ly.david.ui.test.screenshot.ScreenshotTest
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(TestParameterInjector::class)
-class ReleaseListScreenTest : ScreenshotTest(isFullScreen = true) {
-
-    @Test
-    fun default() {
-        snapshot {
-            PreviewReleasesListScreen()
-        }
-    }
-}
+// @RunWith(TestParameterInjector::class)
+// class ReleaseListScreenTest : ScreenshotTest(isFullScreen = true) {
+//
+//    @Test
+//    fun default() {
+//        snapshot {
+//            PreviewReleasesListScreen()
+//        }
+//    }
+// }
