+++ ui/common/src/androidMain/kotlin/ly/david/ui/common/releasegroup/ReleaseGroupListItemPreview.kt
@@ -1,0 +1,25 @@
+package ly.david.ui.common.releasegroup
+
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import ly.david.musicsearch.core.models.listitem.ReleaseGroupListItemModel
+import ly.david.ui.core.preview.DefaultPreviews
+import ly.david.ui.core.theme.PreviewTheme
+
+private val testReleaseGroup = ReleaseGroupListItemModel(
+    id = "6825ace2-3563-4ac5-8d85-c7bf1334bd2c",
+    name = "欠けた心象、世のよすが",
+    primaryType = "EP",
+    firstReleaseDate = "2021-09-08",
+    formattedArtistCredits = "Some artist feat. some other artist",
+)
+
+@DefaultPreviews
+@Composable
+private fun Preview() {
+    PreviewTheme {
+        Surface {
+            ReleaseGroupListItem(testReleaseGroup)
+        }
+    }
+}
